feat(bootstrap): change nav elements to attributes
