+++ packages-adapters/bootstrap/src/components/dropdown/dropdown-item.ts
@@ -1,10 +1,9 @@
 import './dropdown.scss';
 
+import { coerceBoolean } from '@ekzo-dev/toolkit';
 import { bindable, customAttribute } from 'aurelia';
 
 import { BaseAttribute } from '../base-attribute';
-
-// import { coerceBoolean } from '@ekzo-dev/toolkit';
 
 export type BsDropdownItemType = 'item' | 'divider' | 'text' | 'header';
 
@@ -16,10 +15,10 @@ export class BsDropdownItem extends BaseAttribute {
   @bindable()
   type: BsDropdownItemType = 'item';
 
-  @bindable()
+  @bindable(coerceBoolean)
   disabled: boolean = false;
 
-  @bindable()
+  @bindable(coerceBoolean)
   active: boolean = false;
 
   binding() {
