+++ packages-adapters/bootstrap/src/components/nav/index.ts
@@ -1,5 +1,3 @@
 export * from './nav';
-export * from './nav-item';
-export * from './nav-link';
 export * from './tab-content';
 export * from './tab-pane';
