--- packages-adapters/bootstrap/src/components/nav/nav-item.html
@@ -1,3 +1,0 @@
-<template class="nav-item">
-  <au-slot></au-slot>
-</template>
