--- packages-adapters/bootstrap/src/components/nav/nav.html
@@ -1,3 +1,0 @@
-<template class.bind="classes">
-  <au-slot></au-slot>
-</template>
