+++ packages-adapters/bootstrap/src/components/nav/nav.ts
@@ -1,26 +1,42 @@
-import template from './nav.html';
-
 import './nav.scss';
 
-import { bindable, customElement, ICustomElementViewModel } from 'aurelia';
+import { coerceBoolean } from '@ekzo-dev/toolkit';
+import { bindable, customAttribute } from 'aurelia';
+
+import { BaseAttribute } from '../base-attribute';
 
 export type NavType = 'tabs' | 'pills';
 export type NavFill = 'fill' | 'justified';
 
-@customElement({
-  name: 'bs-nav',
-  template,
-})
-export class BsNav implements ICustomElementViewModel {
+@customAttribute('bs-nav')
+export class BsNav extends BaseAttribute {
   @bindable()
   type?: NavType;
 
   @bindable()
   fill?: NavFill;
 
-  get classes(): string {
-    return ['nav', this.type ? `nav-${this.type}` : false, this.fill ? `nav-${this.fill}` : false]
-      .filter(Boolean)
-      .join(' ');
+  get classes(): string[] {
+    return ['nav', this.type ? `nav-${this.type}` : '', this.fill ? `nav-${this.fill}` : ''].filter(Boolean);
+  }
+}
+
+@customAttribute('bs-nav-item')
+export class BsNavItem extends BaseAttribute {
+  get classes(): string[] {
+    return ['nav-item'];
+  }
+}
+
+@customAttribute('bs-nav-link')
+export class BsNavLink extends BaseAttribute {
+  @bindable(coerceBoolean)
+  active: boolean = false;
+
+  @bindable(coerceBoolean)
+  disabled: boolean = false;
+
+  get classes(): string[] {
+    return ['nav-link', this.active ? 'active' : null, this.disabled ? 'disabled' : null].filter(Boolean);
   }
 }
