--- packages-adapters/bootstrap/src/components/navbar/navbar-nav.html
@@ -1,3 +1,0 @@
-<template class="navbar-nav ${scroll ? 'navbar-nav-scroll' : ''}">
-  <au-slot></au-slot>
-</template>
